se agrega consulta detalle de actor por cveActorAsunto

+++ src/main/java/mx/gob/imss/cit/mjlssc/model/request/AbogadoResponsableDto.java
@@ -20,7 +20,7 @@ public class AbogadoResponsableDto implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = 128052803305186026L;
-	private Long abogado;
+	private Integer abogado;
 	private Date fechaAsignacion;
 	
 }
